refactor: package 정리 (#55)
