+++ src/components/PC/Gear/Detail/GearSetting/GearSetting.jsx
@@ -12,8 +12,8 @@ import {
     TextField,
     Typography,
 } from "@mui/material";
-import { ActionButtons } from "../ActionButtons/ActionButtons";
-import { SectionTemplate } from "../SectionTemplate";
+import { ActionButtons } from "../common/ActionButtons";
+import { SectionTemplate } from "../common/SectionTemplate";
 
 export const GearSetting = () => {
     const models = [
