+++ src/components/PC/Gear/Detail/common/ActionButtons/index.js
