+++ src/components/PC/Gear/Detail/common/SectionTemplate/Section.template.jsx
