+++ src/components/PC/Gear/Detail/common/SectionTemplate/SectionTemplate.stories.mdx
