+++ src/components/PC/Gear/Detail/common/SectionTemplate/index.js
